Added reward description for delegate and delegator
