+++ src/Components/AccountSummary/AccountSummaryHome/Blocks/Status/ToolTipTexts.js
